change val to a float to prevent all blue error

+++ lidarFullW_Alpha/src/LidarVolume.hpp
@@ -17,16 +17,16 @@
 class LidarVolume{
 
   public:
-    //The padded min and max fields describing the bounding box(bb) that 
-    //includes the first & last points of the sampled parts of the returning 
-    //waveforms of all pulses. 
+    //The padded min and max fields describing the bounding box(bb) that
+    //includes the first & last points of the sampled parts of the returning
+    //waveforms of all pulses.
     double bb_x_min_padded;
     double bb_y_min_padded;
     double bb_z_min_padded;
     double bb_x_max_padded;
     double bb_y_max_padded;
     double bb_z_max_padded;
- 
+
     int bb_i_min;
     int bb_j_min;
     int bb_k_min;
@@ -41,18 +41,18 @@ class LidarVolume{
     int i_extent;
     int j_extent;
     int k_extent;
-    
-    //number of peaks to be saved into 
+
+    //number of peaks to be saved into
     int numOfPeaks;
-    
+
     //current position in peaks array
-    int currentPeak;    
-   
+    int currentPeak;
+
     //The raw raster data
     int* raster;
-    
+
     std::vector<Peak>** volume;
-    
+
     LidarVolume();
 
     //Read and store the mins and maxes from the header, calculate and store the
@@ -73,14 +73,14 @@ class LidarVolume{
 
     int toPng(std::string filename);
 
-    // Creates a test image for saving. Creates a Mandelbrot Set fractal of 
+    // Creates a test image for saving. Creates a Mandelbrot Set fractal of
     // size width x height
     float *createImage(int width, int height);
 
-    // This takes the float value 'val', converts it to red, green & 
-    // blue values, then sets those values into the image memory buffer 
+    // This takes the float value 'val', converts it to red, green &
+    // blue values, then sets those values into the image memory buffer
     // location pointed to by 'ptr'
-    void setRGB(unsigned char* r,unsigned char* g, unsigned char* b, int val);
+    void setRGB(unsigned char* r,unsigned char* g, unsigned char* b, float val);
 
     int writeImage(const char* filename, const char* title);
